idm-grpc-service/requirements.txt
grpcio
grpcio-tools
pyjwt
bcrypt
python-dotenv
